BEAD DRIVE 20170308_bx(Ver_161018.8)

+++ BEAD DRIVE_code/SHEET SHUTTER DRIVE/Application/Application.c
@@ -399,7 +399,9 @@ VOID updateDriveFaultFlags(VOID)
 		}
 	}
 
-	if(uDriveApplBlockEEP.stEEPDriveApplBlock.operationCount_A600 >= (uDriveApplBlockEEP.stEEPDriveApplBlock.maintenanceCountLimit_A025*1000))
+// 2017/3/8 Measures against mistake maintenanceCountOverflow. by IME
+//	if(uDriveApplBlockEEP.stEEPDriveApplBlock.operationCount_A600 >= (uDriveApplBlockEEP.stEEPDriveApplBlock.maintenanceCountLimit_A025*1000))
+	if(uDriveApplBlockEEP.stEEPDriveApplBlock.operationCount_A600 >= ((UINT32)uDriveApplBlockEEP.stEEPDriveApplBlock.maintenanceCountLimit_A025*1000))
 	{
 		uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveApplicationFault.bits.maintenanceCountOverflow = TRUE;
 	}
@@ -601,7 +603,7 @@ VOID updateDriveStatusFlags(VOID)
 
         if((rampOutputStatus.shutterCurrentPosition < (uDriveCommonBlockEEP.stEEPDriveCommonBlock.lowerStoppingPos_A101 -    //bug_NO.12?13?14?15
             uDriveApplBlockEEP.stEEPDriveApplBlock.overrunProtection_A112)) && (rampOutputStatus.shutterCurrentPosition >
-            (uDriveCommonBlockEEP.stEEPDriveCommonBlock.apertureHeightPos_A130 +    
+            (uDriveCommonBlockEEP.stEEPDriveCommonBlock.apertureHeightPos_A130 +
             uDriveApplBlockEEP.stEEPDriveApplBlock.overrunProtection_A112)))
         {
             uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveStatus.bits.shutterBetweenLowlmtAphgt = TRUE;
@@ -699,7 +701,9 @@ VOID updateSytemCounters(VOID)
 				{
 					// if aperture count exceeds the EEPROM parameter -- then trigger a runtime calibration operation,
 					// at the end of runtime calibration, the shutter should come to rest at the aperture height instead of upper limit
-					if(uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)
+// 2017/3/8 by IME
+//					if(uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)
+					if((UINT32)uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)
 					{
                         uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveStatus.bits.driveReady = FALSE;
                         uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveApplicationFault.bits.driveCalibrationFailed = FALSE;
@@ -760,7 +764,9 @@ VOID updateSytemCounters(VOID)
 				{
 					// if aperture count exceeds the EEPROM parameter -- then trigger a runtime calibration operation,
 					// at the end of runtime calibration, the shutter should come to rest at the aperture height instead of upper limit
-					if(uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)
+// 2017/3/8 by IME
+//					if(uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)
+					if((UINT32)uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)
 					{
                         uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveStatus.bits.driveReady = FALSE;
                         uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveApplicationFault.bits.driveCalibrationFailed = FALSE;
@@ -833,7 +839,9 @@ VOID updateSytemCounters(VOID)
 				{
 					// if aperture count exceeds the EEPROM parameter -- then trigger a runtime calibration operation,
 					// at the end of runtime calibration, the shutter should come to rest at the aperture height instead of upper limit
-					if((uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 - 1) <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)    //bug_NO.12
+// 2017/3/8 by IME
+//					if((uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 - 1) <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)    //bug_NO.12
+					if(((UINT32)uDriveApplBlockEEP.stEEPDriveApplBlock.correctedFreqAperture_A126 - 1) <= uDriveApplBlockEEP.stEEPDriveApplBlock.apertureHeightOperCount_A604)    //bug_NO.12
 					{
 						//	Added to handle aperture correction when operation count is incremented at upper limit - Jan 2016
 //						lsbStartApertureCorrection = TRUE; //20161202
@@ -872,7 +880,7 @@ VOID updateSytemCounters(VOID)
 		sShutterAtUpperLimitFlag = FALSE;
     }
 
-	else if(bUpApertureCmdRecd && !uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveStatus.bits.shutterApertureHeight)  //20161201 
+	else if(bUpApertureCmdRecd && !uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveStatus.bits.shutterApertureHeight)  //20161201
 //	else if(bDownApertureCmdRecd && !uDriveStatusFaultBlockEEP.stEEPDriveStatFaultBlock.uDriveStatus.bits.shutterApertureHeight)
     {
         incrementOperationCnt = TRUE;
